otros constructores, construcctor con opcion

+++ sistema_notas/src/main/java/co/edu/utp/missiontic2022/c2/Nota.java
@@ -71,7 +71,7 @@ public class Nota {
         //this.cualitativo = pEscala100 >= 60 ? "Aprobado" : "Reprobado";
     }
 
-    Nota(String pNombre, int pEscala5){ 
+    Nota(String pNombre, double pEscala5){ 
         this.nombre = pNombre;
         this.escala5 = pEscala5;
         if(pEscala5 >= 2.95){
